+++ DKey.Algorithms.Tests/Graph-likeStructures/BinaryTreeTests.cs
@@ -4,8 +4,6 @@ namespace DKey.Algorithms.Tests.Graph_likeStructures;
 
 public class IntegerSumBinaryTree : ImmutableBinaryTree<int>
 {
-    public IntegerSumBinaryTree(IList<int> data) : base(data) { }
-
     protected override int Add(int a, int b)
     {
         return a + b;
@@ -24,11 +22,11 @@ public class BinaryTreeTests
     public void GetSumTest()
     {
         var data = new List<int> { 1, 2, 3, 4, 5 };
-        var tree = new IntegerSumBinaryTree(data);
+        var tree = new IntegerSumBinaryTree();
 
-        Assert.AreEqual(15, tree.GetSum(0, 4));
-        Assert.AreEqual(6, tree.GetSum(0, 2));
-        Assert.AreEqual(7, tree.GetSum(2, 3));
-        Assert.AreEqual(9, tree.GetSum(3, 4));
+        Assert.AreEqual(15, tree.GetCumulativeOperation(0, 4));
+        Assert.AreEqual(6, tree.GetCumulativeOperation(0, 2));
+        Assert.AreEqual(7, tree.GetCumulativeOperation(2, 3));
+        Assert.AreEqual(9, tree.GetCumulativeOperation(3, 4));
     }
 }