+++ DKey.CodeForces/Program.cs
@@ -1,11 +1,11 @@
 using DKey.Algorithms;
-using DKey.CodeForces.Contest1816;
+using DKey.CodeForces.Contest1955;
 
 namespace DKey.CodeForces;
 
 public static class Program
 {
-    private static Solver _solver = new Solver1816E();
+    private static Solver _solver = new Solver1955H();
     public static void Main()
     {
         _solver.Run();