+++ DKey.ContestTemplateBuilder/Config.cs
@@ -2,6 +2,6 @@
 
 public class Config
 {
-    public static string repoPath = @"C:\Repos\github\DKey.Algorithms\DKey.CodeForces";
-    public static string contestID = "1816";
+    public static string repoPath = @"C:\Repos\GitHub\DKey.Algorithms\DKey.CodeForces";
+    public static string contestID = "1955";
 }